navigation-backend/requirements.txt
fastapi
uvicorn
requests
streamlit
pydantic-settings
python-dotenv
pillow
python-multipart